+++ KnapsackProblem.Common/KnapsackLoader.cs
@@ -16,7 +16,7 @@ namespace KnapsackProblem.Common
         private static Knapsack ParseKnapsack(string line)
         {
             var parts = line.Split();
-            return new Knapsack(short.Parse(parts[0]), short.Parse(parts[2]), ParseKnapsackItems(parts.Skip(3)));
+            return new Knapsack(int.Parse(parts[0]), int.Parse(parts[2]), ParseKnapsackItems(parts.Skip(3)));
         }
 
         private static IEnumerable<KnapsackItem> ParseKnapsackItems(IEnumerable<string> enumerable)
